refactor(i18n-loader): Extract code for loading i18n content.
