+++ lib/actions/ui.js
@@ -6,8 +6,7 @@ import coreUtils from '@opentripplanner/core-utils'
 import {
   getConfigLocales,
   getDefaultLocale,
-  getMatchingLocaleString,
-  loadLocaleData
+  getMatchingLocaleString
 } from '../util/i18n'
 import {
   getItineraryView,
@@ -18,6 +17,7 @@ import {
   ItineraryView
 } from '../util/ui'
 import { getModesForActiveAgencyFilter, getUiUrlParams } from '../util/state'
+import { loadLocaleData } from '../util/i18n-loader'
 
 import {
   clearActiveSearch,
