+++ lib/util/i18n.js
@@ -1,10 +1,4 @@
-import flatten from 'flat'
 import memoize from 'lodash.memoize'
-import merge from 'deepmerge'
-
-// Vite-specific glob import of all i18n YML content in OTP-RR and @opentripplanner packages.
-const i18nModules = import.meta.glob('../../i18n/*.yml', { import: 'default' })
-const i18nOtpUiModules = import.meta.glob('../../node_modules/@opentripplanner/**/i18n/*.yml', { import: 'default' })
 
 /**
  * Match the provided locale (language and region) to the list of supported locales.
@@ -78,47 +72,6 @@ export const getLanguageOptions = memoize((configLanguages) => {
   })
   return filteredLanguages
 })
-
-/**
- * Loads and merge all localized strings from known @opentripplanner packages that have translations.
- */
-async function loadOtpUiLocaleData(matchedLocale) {
-  const fileEnding = `/i18n/${matchedLocale}.yml`
-  const otpUiMessages = await Promise.all(
-    // Retain OTP-UI i18n content for the specified locale
-    Object.keys(i18nOtpUiModules)
-      .filter((key) => key.endsWith(fileEnding))
-      .map(async (key) => i18nOtpUiModules[key]())
-  )
-  return merge.all(otpUiMessages)
-}
-
-/*
- * Load the localized strings for all languages here,
- * merge with the provided customized strings from state.otp.config.language
- * (overrides can be defined per language or for all languages)
- * and flatten to a map of
- *   id => string.
- * FIXME: Load languages on demand using fetch.
- */
-export async function loadLocaleData(matchedLocale, customMessages) {
-  // If only 'zh' is configured and no other Chinese subset is, use Chinese Simplified.
-  const loadedLocale = matchedLocale === 'zh' ? 'zh-Hans' : matchedLocale
-
-  const messages = await i18nModules[`../../i18n/${loadedLocale}.yml`]()
-  const otpUiMessages = await loadOtpUiLocaleData(loadedLocale)
-
-  // Merge custom strings into the standard language strings.
-  const mergedMessages = {
-    ...flatten(otpUiMessages),
-    ...flatten(messages),
-    // Override the predefined strings with the custom ones, if any provided.
-    ...flatten(customMessages.allLanguages || {}),
-    ...flatten(customMessages[matchedLocale] || {})
-  }
-
-  return mergedMessages
-}
 
 /**
  * Gets the localization > defaultLocale configuration setting, or 'en-US' if not set.
